add alert for submit and delete action

+++ src/components/FeedbackItem/FeedbackItem.js
@@ -3,10 +3,10 @@ import React, { Component } from 'react';
 import { makeStyles } from '@material-ui/core/styles';
 import TableCell from '@material-ui/core/TableCell';
 
-import DeleteIcon from '@material-ui/icons/Delete';
 import IconButton from '@material-ui/core/IconButton';
+import DeleteIcon from '@material-ui/icons/Delete';
 
-
+import Swal from 'sweetalert2'
 
 class FeedbackItem extends Component {
     //Styling with Material-UI
@@ -24,11 +24,49 @@ class FeedbackItem extends Component {
    
     //End styling with Material-UI
 
+//handle delete feedback with SweetAlert2
     handleDeleteBtn = () => {
-        console.log('delete btn clicked!')
-        console.log('id of feedback to delete', this.props.feedback.id);
-        this.props.deleteFeedback(this.props.feedback.id)
+        const swalWithBootstrapButtons = Swal.mixin({
+            customClass: {
+                confirmButton: 'btn btn-success',
+                cancelButton: 'btn btn-danger'
+            },
+            buttonsStyling: false
+        })
+
+        swalWithBootstrapButtons.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonText: 'Yes, delete it!',
+            cancelButtonText: 'No, cancel!',
+            reverseButtons: true
+        }).then((result) => {
+            if (result.value) {
+
+                console.log('delete btn clicked!')
+                console.log('id of feedback to delete', this.props.feedback.id);
+                this.props.deleteFeedback(this.props.feedback.id);
+
+                swalWithBootstrapButtons.fire(
+                    'Deleted!',
+                    'This feedback has been deleted.',
+                    'success'
+                )
+            } else if (
+                /* Read more about handling dismissals below */
+                result.dismiss === Swal.DismissReason.cancel
+            ) {
+                swalWithBootstrapButtons.fire(
+                    'Cancelled',
+                    'This feedback is safe.',
+                    'error',
+                )
+            }
+        })
     }
+
 
     render() {
         const classes = this.useStyles;
@@ -39,7 +77,6 @@ class FeedbackItem extends Component {
                 <TableCell >{this.props.feedback.understanding}</TableCell>
                 <TableCell >{this.props.feedback.support}</TableCell>
                 <TableCell >{this.props.feedback.comments}</TableCell>
-                {/* <TableCell ><button onClick={() => this.handleDeleteBtn(this.props.feedback.id)}>Delete</button></TableCell> */}
 
                 <TableCell >
                     <IconButton
